Added new query and view controls

+++ CatAlog_App.Db/DbContexts/DbModels/VideoModels/SerialData.cs
@@ -5,18 +5,18 @@ namespace CatAlog_App.Db.DbContexts.DbModels.VideoModels
 {
     public class SerialData
     {
-        public int Id { get; set; }
+        public uint Id { get; set; }
 
-        public int MainRecordDataId { get; set; }
+        public uint MainRecordDataId { get; set; }
         public MainRecordData MainRecordData { get; set; }
 
         public IEnumerable<EpisodeData> Episodes { get; set; }
 
-        public int? SeasonNumber { get; set; }
+        public byte? SeasonNumber { get; set; }
 
         public string Type { get; set; }
 
-        public int? CountSpecials { get; set; }
+        public byte? CountSpecials { get; set; }
 
         public string Note { get; set; }
     }
