+++ CatAlog_App.Db/DtoModels/DtoCatalogHierarchy.cs
@@ -13,6 +13,6 @@ namespace CatAlog_App.Db.DtoModels
     {
         public string CategoryName { get; set; }
 
-        public int Count { get; set; }
+        public ushort Count { get; set; }
     }
 }
